refine upload & parse (#1969)
### What problem does this PR solve?


### Type of change
- [x] Refactoring
